resolved comments and updated logic for cvt_archive
