+++ ribs/archives/_cvt_archive.py
@@ -83,17 +83,21 @@ class CVTArchive(ArchiveBase):
         dtype (str or data-type): Data type of the solutions, objectives,
             and measures. We only support ``"f"`` / ``np.float32`` and ``"d"`` /
             ``np.float64``.
+        custom_centroids (array-like): If passed in, this (cells, measure_dim)
+            array will be used as the centroids of the CVT instead of generating
+            new ones. In this case, ``samples`` will be ignored, and
+            ``archive.samples`` will be None. This can be useful when one wishes
+            to use the same CVT across experiments for fair comparison.
+        centroid_method (string): Pass in the following methods for
+            generating centroids: "random", "sobol", "scrambled sobol",
+            "halton". Default method is "kmeans". Note: Samples are only used
+            when method is "kmeans".
         samples (int or array-like): If it is an int, this specifies the number
             of samples to generate when creating the CVT. Otherwise, this must
             be a (num_samples, measure_dim) array where samples[i] is a sample
             to use when creating the CVT. It can be useful to pass in custom
             samples when there are restrictions on what samples in the measure
             space are (physically) possible.
-        custom_centroids (array-like): If passed in, this (cells, measure_dim)
-            array will be used as the centroids of the CVT instead of generating
-            new ones. In this case, ``samples`` will be ignored, and
-            ``archive.samples`` will be None. This can be useful when one wishes
-            to use the same CVT across experiments for fair comparison.
         k_means_kwargs (dict): kwargs for :func:`~sklearn.cluster.k_means`. By
             default, we pass in `n_init=1`, `init="random"`,
             `algorithm="lloyd"`, and `random_state=seed`.
@@ -120,13 +124,13 @@ class CVTArchive(ArchiveBase):
                  qd_score_offset=0.0,
                  seed=None,
                  dtype=np.float64,
-                 samples=100_000,
                  custom_centroids=None,
-                 chunk_size=None,
+                 centroid_method="kmeans",
+                 samples=100_000,
                  k_means_kwargs=None,
                  use_kd_tree=True,
                  ckdtree_kwargs=None,
-                 centroid_method="kmeans"):
+                 chunk_size=None):
 
         ArchiveBase.__init__(
             self,
@@ -167,18 +171,17 @@ class CVTArchive(ArchiveBase):
 
         if custom_centroids is None:
             if centroid_method == "kmeans":
+                # Samples are set to custom first and then checked or used.
+                self._samples = samples
                 if not isinstance(samples, int):
-                    # Validate shape of custom samples. These are ignored when
-                    # `custom_centroids` is provided.
+                    # Validate shape of custom samples.
                     samples = np.asarray(samples, dtype=self.dtype)
                     if samples.shape[1] != self._measure_dim:
                         raise ValueError(
                             f"Samples has shape {samples.shape} but must be of "
                             f"shape (n_samples, len(ranges)="
                             f"{self._measure_dim})")
-                self._samples = samples
-                self._centroids = None
-                if self._centroids is None:
+                else:
                     self._samples = self._rng.uniform(
                         self._lower_bounds,
                         self._upper_bounds,
@@ -186,34 +189,39 @@ class CVTArchive(ArchiveBase):
                     ).astype(self.dtype) if isinstance(self._samples,
                                                        int) else self._samples
 
-                    self._centroids = k_means(self._samples, self._cells,
-                                              **self._k_means_kwargs)[0]
+                self._centroids = k_means(self._samples, self._cells,
+                                          **self._k_means_kwargs)[0]
 
-                    if self._centroids.shape[0] < self._cells:
-                        raise RuntimeError(
-                            "While generating the CVT, k-means clustering "
-                            f"found {self._centroids.shape[0]} centroids, "
-                            f"but this archive needs {self._cells} cells. "
-                            f"This most likely happened because there are "
-                            "too few samples and/or too many cells.")
+                if self._centroids.shape[0] < self._cells:
+                    raise RuntimeError(
+                        "While generating the CVT, k-means clustering found "
+                        f"{self._centroids.shape[0]} centroids, but this "
+                        f"archive needs {self._cells} cells. This most "
+                        "likely happened because there are too few samples "
+                        "and/or too many cells.")
+
+                if self._use_kd_tree:
+                    self._centroid_kd_tree = cKDTree(self._centroids,
+                                                     **self._ckdtree_kwargs)
             elif centroid_method == "random":
-                # generates random centroids across two ranges
+                # Generate random centroids for the archive.
                 self._centroids = self._rng.uniform(self._lower_bounds,
                                                     self._upper_bounds,
-                                                    size=(self._cells, 2))
+                                                    size=(self._cells,
+                                                          self._measure_dim))
             elif centroid_method == "sobol":
-                # generates Sobol numbers
-                sampler = Sobol(d=2, scramble=False)
+                # Generate self._cells number of centroids as a Sobol sequence.
+                sampler = Sobol(d=self._measure_dim, scramble=False)
                 num_points = np.log2(self._cells).astype(int)
                 self._centroids = sampler.random_base2(num_points)
-            elif centroid_method == "scrambled sobol":
-                # generates scrambled Sobol numbers
-                sampler = Sobol(d=2, scramble=True)
+            elif centroid_method == "scrambled_sobol":
+                # Generates centroids as a scrambled Sobol sequence.
+                sampler = Sobol(d=self._measure_dim, scramble=True)
                 num_points = np.log2(self._cells).astype(int)
                 self._centroids = sampler.random_base2(num_points)
             elif centroid_method == "halton":
-                # generates Halton numbers
-                sampler = Halton(d=2)
+                # Generates centroids using a Halton sequence.
+                sampler = Halton(d=self._measure_dim)
                 self._centroids = sampler.random(n=self._cells)
         else:
             # Validate shape of `custom_centroids` when they are provided.
@@ -225,10 +233,6 @@ class CVTArchive(ArchiveBase):
                     f"{self._measure_dim})")
             self._centroids = custom_centroids
             self._samples = None
-
-        if self._use_kd_tree:
-            self._centroid_kd_tree = cKDTree(self._centroids,
-                                             **self._ckdtree_kwargs)
 
     @property
     def lower_bounds(self):
